refactor: Move functions to BlockManager methods
This is a commit in preparation for the next three commits. The
functions are moved to methods to avoid their re-declaration for the
purpose of passing in BlockManager options.

To allow for the move of functions to BlockManager methods all call
sites require an instantiated BlockManager, or a callback to one.