refactor, blockstorage: Replace fastprune arg
The BlockManager class now holds a fastprune option that can be used by
other functions in blockstorage.* instead of the global gArgs. To
this end move functions currently relying on gArgs into the blockstorage
class, or add a function argument to allow calling the function
without gArgs in scope.

This should eventually allow users of the BlockManager to not rely on
the global Args and instead pass in their own options.