refactor, blockstorage: Replace blocksdir arg
Add a blocks_dir field to the BlockManager options. Move functions
relying on the global gArgs to get the blocks_dir into the BlockManager
class, or add blocks_dir function arguments to use the blocks_dir from
the options instead of gArgs.

This should eventually allow users of the BlockManager to not rely on
the global Args and instead pass in their own options.