refactor: Move chain names to the kernel namespace
The commit is part of an ongoing effort to decouple the libbitcoinkernel
library from the ArgsManager and other functionality that should not be
part of the kernel library.

The code move of the chain name constants out of the chainparamsbase to
their own separate header allows the kernel chainparams to no longer
include chainparamsbase. The chainparamsbase contain references to
the ArgsManager and networking related options.