refactor, blockstorage: Replace fastprune arg
The BlockManager class now holds a fastprune field in its options that
can be used by other functions in blockstorage.* instead of the global
gArgs. To this end move functions currently relying on gArgs into the
blockstorage class.

This should eventually allow users of the BlockManager to not rely on
the global Args and instead pass in their own options.