refactor: Create chainname files
This is the first of a number of commits with the goal of moving the
chain name definitions out of chainparamsbase to their own file. The
goal is to allow the kernel chainparams to no longer include
chainparamsbase.

The commit is part of an ongoing effort to decouple the libbitcoinkernel
library from the ArgsManager and other functionality that should not be
part of the kernel library.