refactor: Extract common/args from util/system
This is an extraction of ArgsManager related functions from util/system
into their own common file.

The background of this commit is an ongoing effort to decouple the
libbitcoinkernel library from the ArgsManager. The ArgsManager belongs
into the common library, since the kernel library should not depend on
it. See doc/design/libraries.md for more information on this rationale.