refactor: Pass FatalError in options to ChainstateManager
FatalError replaces what previously was the AbortNode function in
shutdown.cpp.

This commit is part of the libbitcoinkernel project and removes the
shutdown's and, more generally, the kernel library's dependency on
interface_ui with an options method. By removing interface_ui from the
kernel library, its dependency on boost is reduced to just
boost::multi_index.