kernel: Add fatalError method to notifications
FatalError replaces what previously was the AbortNode function in
shutdown.cpp.

This commit is part of the libbitcoinkernel project and removes the
shutdown's and, more generally, the kernel library's dependency on
interface_ui with an options method. By removing interface_ui from the
kernel library, its dependency on boost is reduced to just
boost::multi_index.

Also add shutdown switch in node::KernelNotifications. This simplifies
testing of validation methods that may normally trigger a shutdown on a
fatal error by leaving out the call to StartShutdown.

Co-authored-by: Russell Yanofsky <russ@yanofsky.org>
Co-authored-by: TheCharlatan <seb.kung@gmail.com>