refactor: Pass BlockNotify in options to ChainstateManager
This commit is part of the libbitcoinkernel project and seeks to remove
the ChainstateManager's and, more generally, the kernel library's
dependency on interface_ui with options methods in this and the following
few commits. By removing interface_ui from the kernel library, its
dependency on boost is reduced to just boost::multi_index.

Create a new file in the node library for defining a function creating
the default set of notification methods such that these do not need to
be re-defined all over the codebase.