kernel: Remove shutdown from kernel library
This decouples the libbitcoinkernel library from the shutdown routines.
As a library, it should not directly rely on code that manages the
termination of processes.

While touching the lines around StartShutdown, it became apparent that
there was no logging indicating the reason for a call to StartShutdown.
Add ShutdownReason enum to the startShutdown notification to communicate
this.