refactor: Split up ThreadImport method
Leave the blockimport part in its current place in blockstorage, call
mempool LoadMempool directly from the load_block thread in init.cpp.

This has the benefit of moving the StartShutdown call out of the
blockstorage file and thus out of the kernel's responsibility. The user
can now decide if he wants to shut down after a block import or not.