refactor: Add read and write lamdbas for CDBWrapper::Read
This wraps manipulating the data streams for reading and writing the db
keys and values in two lambdas such that these can be used without
specifying template arguments by the dbwrapper implementation.

The context of this commit is an effort to decouple the dbwrapper header
file from leveldb includes. To this end, the includes are moved to the
dbwrapper implementation file. This is done as part of the kernel
project to reduce the number of required includes for users of the
kernel.

Co-authored-by: MarcoFalke <*~=`'#}+{/-|&$^_@721217.xyz>