blockstorage: Return on fatal block file flush error
By returning if `FlushBlockFile` failed, we ensure that no further write
operations take place that may lead to an inconsistent database when
crashing. Add a [[nodiscard]] annotation such that it is not ignored in
future.

Prior to this patch `FlushBlockFile` may have failed without returning
in `Chainstate::FlushStateToDisk`, leading to a potential write from
`WriteBlockIndexDB` that may refer to a block that is not fully
flushed to disk yet.

Functions that have `FlushBlockFile` in their call stack, need to handle
this abort case properly. Since `Chainstate::FlushStateToDisk` already
produces an abort error in case of its call to `WriteBlockIndexDB`
failing, no extra logic for functions calling
`Chainstate::FlushStateToDisk` is required.

Besides `Chainstate::FlushStateToDisk`, `FlushBlockFile` is also called
by `FindBlockPos`. Don't change the abort behavior there, since we don't
want to fail the function if the flushing of already written blocks
fails. Instead, just document it.