blockstorage: Return on fatal undo file flush error
By returning if either `FlushUndoFile` or `FlushBlockFile` fail, we
ensure that no further write operations take place that may lead to an
inconsistent database when crashing.

Besides the call to `FlushUndoFile` in `FlushBlockFile`, ignore its
return code at its call site in `WriteUndoDataForBlock`. There, a failed
flush of the undo data should not be indicative of a failed write.

Add [[nodiscard]] annotations to `FlushUndoFile` such that its return
value is not ignored in future.