[refactor] Make txmempool's GetSortedDepthAndScore static
This removes the need for defining its complex boost multi index return
type in the header.

The context of this change is an effort towards removing the boost multi
index includes from the header tree. The goal is that the experimental
`bitcoin-chainstate` binary can be compiled without requiring boost
includes.