[refactor] Move CTxMemPool::visited implementation to implementation file
This allows the function to use txiter without defining its full
implementation manipulating a complex boost multi index type in the
header.

The context of this change is an effort towards removing the boost multi
index includes from the header tree. The goal is that the experimental
`bitcoin-chainstate` binary can be compiled without requiring boost
includes.