build: Prepare libbitcoin_util for internal kernel library
Move util library files that are not used by the kernel library and are
not unopinionated utilities to the common library. This is consistent
with the library structure laid out in doc/design/libraries.md.

Moving util/message.cpp to the common library also breaks an
undocumented dependency of the util library on the consensus library's
CPubKey::RecoverCompact symbol.