kernel: De-globalize fReindex
fReindex is one of the last remaining globals exposed by the kernel
library, so move it into the blockstorage class to reduce the amount of
global mutable state and make the kernel library a bit less awkward to
use.

This commit also slightly changes behaviour around restarting a node
busy with a reindex without passing the "-reindex" argument again.
Before, this would leave the block and coins data intact, but discard
the data of the optional indexes. While not a bug per se, since no
relevant validation signals are issued for the index while the
reindexing atomic is persisted as true (and thus the reindex continued
on restart), it is brittle and not consistent with the behaviour of the
other indexes. Align this by only wiping the optional indexes when the
"-reindex" argument is provided.