[refactor] Simplify setting load chainstate reindex option
The reindex option for ChainstateLoadOptions is currently set in a
confusing way. By using the reindex atomic, it is not obvious in which
scenario it is true or false.

The atomic is controlled by both the user passing the -reindex option,
and by reading the reindexing flag from the block tree database in
LoadBlockIndexDB. In practice this read is done through the chainstate
module's CompleteChainstateInitialization's call to LoadBlockIndex.
Since this is only done after the reindex option is set already, it does
not have an effect on it.

Make this clear by using the reindex option from the blockman opts which
is only controlled by the user.