kerenl: Add chain params context option to C header
As a first option, add the chainparams. For now these can only be
instantiated with default values. In future they may be expanded to take
their own options for regtest and signet configurations.

This commit also introduces a unique pattern for setting the option
values. Since the number of potential future options that may be
configured is not really bounded, it may hurt discoverability in the
header to just add a function for each of them. Instead add a single
function for setting the options of an option object, in this case
`kernel_context_options_set(...)`, that takes a unique identifier and a
void* value as arguments. The user passes the option value to be
configured as a void pointer alongside its identifier. Valid option
values are tagged internally in their struct and checked if their value
matches with the selected identifier when calling the `*_set(...)`
function.