Introduce internal kernel library
Do not compile the kernel library if reduced exports are switched on.
Symbols from the kernel library won't be reduced anyway, so there is no
point in doing so in combination. Leave the option there will break
compilation currently, since no symbols will be exported from the
dependent libraries of the kernel library and thus make no symbols
available for bitcoin-chainstate.