init: Use size_t consistently for cache sizes
This avoids having to rely on implicit casts when passing them to the
various functions allocating the caches.

This also fixes a bug where an incorrect amount of memory was allocated
on 32 bit platforms if the user configures more than 4000MiB of
-dbcache. If the db cache size exceeded the maximum size of a 32 bit
unsigned integer, it would overflow and allocate much less memory than
instructed to. Similar things would happen if the user configures an
unreasonably high amount of -dbcache on 64 bit platforms.

Fix this by clamping the total amount of allocated db_cache to the
maximum value of a size_t if a too high -dbcache is passed in.

Also take this opportunity to make the total amounts of cache in the
chainstate manager a size_t too.