kernel: make ChainstateManagerOptions use kernel_LockedDirectory
Note: to ensure that data_dir is kept alive for the duration of
ChainstateManager, it is added as a parameter to
kernel_chainstate_manager_destroy. This introduces some asymmetry
where data_dir is a parameter to destroy, but not create,
ChainstateManager. This could be addressed by refactoring data_dir
out of ChainstateManagerOptions and adding it as a ctor parameter
to ChainstateManager, but is not done here to keep the changes
minimal.